src/requirements.txt
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
python-dotenv==0.10.3
six==1.13.0
SQLAlchemy==1.3.12
wcwidth==0.1.7
Werkzeug==0.16.0
